Update style.css
so i managed to at least make a triangle and turn it around, but i ran out of time today before i could figure out how to arrange and stack the elements, and then I will have the challenge of getting the nose shape to copy one of the color sides. But that will be a good learning oppurtunity for the week!